requirements.txt
PyQt6>=6.7.0
Pillow>=10.0.0
piexif>=1.1.3
exifread>=3.0.0
rawpy>=0.18.0
numpy>=1.24.0
pytest>=8.0.0
pytest-qt>=4.4.0
